Fixes for facets 2.7.
Some required files were removed from facets 2.7. I was accidentally
pulling in any version greater than 2.4.4, which is was the code was
developed against.

Rather than continue with an old version, I upgraded the code to use 2.7
and changed the version strings to use a less optimistic check (~>
instead of >=).