Added amqp-spy. It allows you to listen in on all messages going through a specific exchange.
It creates a temporary auto-delete queue, binds it
to the exchange in question with an optional
routing key, and subscribes to the temp queue.

It's usage isn't perfect, cause you need to match
the exchange's type, durability and auto-delete
options manually rather than amqp-spy figuring out
the values on it's own. But it works, and is very
useful :)